# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: MyCargoLens Public API
  version: "1.0.0-foundation"
  description: >
    Customer-facing API for brokers, 3PLs, and ERP integrations. Authenticate
    with an API key (issued in the app under Settings → API). Requests run under
    the key's organization and its subscription-tier capabilities — the API can
    only do what the org's plan unlocks.


    Status: authentication + read endpoints + ISF write (create draft, submit to
    CBP) + ABI Entry create. ABI Entry send, XML content-negotiation, and status
    webhooks are coming next.
servers:
  - url: https://app.mycargolens.com/api/public/v1
    description: Production
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: mcl_live_*
      description: >
        Send your key as `Authorization: Bearer mcl_live_…` (or the
        `X-API-Key` header). Keys are scoped (filings:read, filings:write,
        entries:read, entries:write) and can be revoked anytime.
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
        code: { type: string, description: "e.g. api_key_required, api_key_invalid, insufficient_scope, feature_not_in_plan" }
    Ping:
      type: object
      properties:
        ok: { type: boolean }
        orgId: { type: string, format: uuid }
        plan: { type: string, nullable: true, description: "active tier id, or null if no active plan" }
        capabilities:
          type: array
          items: { type: string, enum: [ISF_FILING, ABI_ENTRY, CONTAINER_TRACKING, HTS_CLASSIFICATION] }
        scopes:
          type: array
          items: { type: string }
    Shipment:
      type: object
      description: An ISF filing (the shipment record).
      properties:
        id: { type: string, format: uuid }
        filingType: { type: string, enum: [ISF-10, ISF-5] }
        status: { type: string, example: submitted }
        masterBol: { type: string, nullable: true }
        houseBol: { type: string, nullable: true }
        ccFilingId: { type: string, nullable: true }
        submittedAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    Entry:
      type: object
      description: An ABI Entry document (7501 / 3461). Requires the ABI_ENTRY capability.
      properties:
        id: { type: string, format: uuid }
        status: { type: string, example: SENT }
        entryType: { type: string, example: "01" }
        entryNumber: { type: string, nullable: true }
        mbolNumber: { type: string, nullable: true }
        sentAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
security:
  - ApiKeyAuth: []
paths:
  /ping:
    get:
      summary: Auth check — returns the org, active plan, capabilities, and the key's scopes.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ping" }
        "401":
          description: Missing or invalid API key
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /shipments:
    post:
      summary: Create an ISF draft (ISF-10 / ISF-5).
      description: Requires scope `filings:write`. Body matches the ISF filing schema (parties, BOLs, commodities, containers). Returns the created draft; it is not submitted to CBP until you call /shipments/{id}/submit.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                filingType: { type: string, enum: [ISF-10, ISF-5], default: ISF-10 }
                masterBol: { type: string }
                houseBol: { type: string }
                commodities: { type: array, items: { type: object } }
                containers: { type: array, items: { type: object } }
              description: "Full schema in server/src/schemas/filing.ts (createFilingSchema)."
      responses:
        "201":
          description: Draft created
          content:
            application/json:
              schema:
                type: object
                properties: { data: { $ref: "#/components/schemas/Shipment" } }
        "400": { description: Validation failed }
        "403": { description: Missing required scope }
    get:
      summary: List ISF filings for the organization.
      description: Requires scope `filings:read`.
      parameters:
        - name: status
          in: query
          schema: { type: string }
          description: Filter by filing status (e.g. draft, submitted, accepted, rejected).
        - name: limit
          in: query
          schema: { type: integer, default: 50, maximum: 200 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Shipment" }
        "401": { description: Missing or invalid API key }
        "403": { description: Missing required scope }
  /shipments/{id}:
    get:
      summary: Get a single ISF filing.
      description: Requires scope `filings:read`.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties: { data: { $ref: "#/components/schemas/Shipment" } }
        "404": { description: Filing not found }
  /shipments/{id}/submit:
    post:
      summary: Submit an ISF filing to CBP.
      description: >
        Requires scope `filings:write` and an active plan (the ISF_FILING
        capability). Bills one charge per shipment on success (idempotent). The
        response mirrors the app: 200 with the updated filing on success, 422 on
        CBP validation errors, 400 on a CBP system error or invalid state, 402
        if no active plan.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Submitted (or transmitted-and-rejected — see filing.status)
          content:
            application/json:
              schema:
                type: object
                properties:
                  filing: { $ref: "#/components/schemas/Shipment" }
                  ccFilingId: { type: string, nullable: true }
        "402": { description: No active plan (code subscription_required) }
        "422": { description: CBP validation errors }
  /entries:
    post:
      summary: Create a DRAFT ABI Entry document (7501 / 3461).
      description: Requires scope `entries:write` and the ABI_ENTRY capability. Optionally pre-fills from a linked ISF filing (`filingId`) or a completed manifest query (`manifestQueryId`). Returns the created draft; transmit it later via /entries/{id}/send (coming next).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                payload: { type: object, description: "ABI entry body (7501/3461 fields)." }
                filingId: { type: string, format: uuid, nullable: true, description: "Pre-fill from a linked ISF filing." }
                manifestQueryId: { type: string, format: uuid, nullable: true, description: "Pre-fill from a manifest query." }
      responses:
        "201":
          description: Draft created
          content:
            application/json:
              schema:
                type: object
                properties: { data: { $ref: "#/components/schemas/Entry" } }
        "400": { description: Validation failed }
        "403": { description: Missing scope or ABI Entry not in plan }
        "404": { description: Linked filing / manifest query not found }
    get:
      summary: List ABI Entry documents for the organization.
      description: Requires scope `entries:read` and the ABI_ENTRY capability (ISF+Entry / Complete tiers).
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 50, maximum: 200 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Entry" }
        "403":
          description: Missing scope, or ABI Entry not in plan (code feature_not_in_plan)
  /entries/{id}:
    get:
      summary: Get a single ABI Entry document.
      description: Requires scope `entries:read` and the ABI_ENTRY capability.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties: { data: { $ref: "#/components/schemas/Entry" } }
        "404": { description: Entry not found }
# ── Coming next release ──────────────────────────────────────────────
# POST /entries/{id}/send      transmit an ABI Entry to CBP   (entries:write + ABI_ENTRY)
# + XML content-negotiation (Accept/Content-Type: application/xml) and status webhooks.